requirements.txt
torch
torchvision
ipython
ipykernel
ipywidgets
numpy
scikit-learn
scikit-image
Pillow
matplotlib
tqdm
opencv-python-headless
urllib3
# urllib3==1.26.6
requests
scipy
ninja
# ===
kornia
open-clip-torch
facenet-pytorch
face-alignment
xformers
# pip install "git+https://github.com/facebookresearch/pytorch3d.git"
yacs
chumpy
# python -m pip install numpy==1.23.1